+++ src/DocumentView.cpp
@@ -142,6 +142,8 @@ DocumentView::initGui() noexcept
     // if (m_config.rendering.icc_color_profile)
     //     m_model->enableICC();
     m_model->setCacheCapacity(m_config.behavior.cache_pages);
+    m_model->setBackgroundColor(m_config.ui.colors.page_background);
+    m_model->setForegroundColor(m_config.ui.colors.page_foreground);
 
     m_hscroll = new ScrollBar(Qt::Horizontal, this);
     m_vscroll = new ScrollBar(Qt::Vertical, this);